refactor: change the home page widget into another file
